requirements.txt
python-dotenv
pandas
numpy
setuptools
pymongo
certifi
pymongo[srv]==3.12
scikit-learn
pyyaml
mlflow
dagshub
fastapi
uvicorn

-e .